+++ martypy/Marty.py
@@ -584,7 +584,7 @@ class Marty(object):
             jointIDNo = self.JOINT_IDS.get(joint_name_or_num, 0)
         return self.client.get_joint_status(jointIDNo)
 
-    def get_distance_sensor(self) -> float:
+    def get_distance_sensor(self) -> int:
         '''
         Get the latest value from the distance sensor :one: :two:
         Returns:
